+++ jobtemplate_test.go
@@ -21,7 +21,7 @@ func TestCreateJobTemplate(t *testing.T) {
 		t.FailNow()
 		return
 	}
-	if job.Height != template.Height {
+	if job.Height != uint64(template.Height) {
 		t.Errorf("job height mismatch: expected %d, got job: %d", template.Height, job.Height)
 	}
 	if job.Subsidy != *template.CoinbaseValue {
